hello world with tab
There are few escape sequences which is used for formatting the output in the python  programming language, you have to use \t escape sequence in order to provide 4 to 5 spaces between words in the output console. Enter the escape sequence directly into the print statement's within the inverted commas ("mes\tsage").